Add files via upload
The Bus Booking System in ASP.NET is a web application enabling users to search and book bus tickets online, offering features like seat selection, payment integration, and booking history. It includes user authentication for customers and admin functionalities for managing routes, schedules, and bookings. Developed using ASP.NET with C# and SQL Server for the backend, it provides a seamless experience for users and efficient management tools for administrators. The system ensures real-time updates and secure transactions for a smooth booking process.